Airball Sensor Board v3
In v2, the Sparkfun Fio v3 was flipped relative to the Airball board,
because of an error stemming from a cleanup of the package. This means
it's not possible to mount the Fio and the pressure sensors at the same
time. :(

In v3, the following things are fixed/changed:
  * Flipped the Sparkfun Fio v3 headers.
  * Rounded the corners of the board a bit more.
  * Updated all silkscreen text to use "vector" font instead of "fixed",
    as that's what the fab actually uses, so it looks like the final
    result in EAGLE.
  * Increased the weight of the pressure sensor labels and a few others
    and done some minor cleanups based on seeing the board in person.

Electrically, v2 was fine and works correctly, so all changes were
basically cosmetic and/or mechanical in nature.