Update airball_probe Arduino sketch to use new library for TMP102
Update the airball_probe sketch to use the newly-written library for the
TI TMP102 temperature sensor, Temperature_LM75_Derived. This should be
functionally the same. A major semantic bug was fixed in that the older
library was reading the high 8 and low 4 bits of the temperature data in
two separate I2C transactions, which could cause corruption.