New firmware for sensor v4; sentences types; API mode
The new data sent from the probe looks like the following examples:

An airdata sentence:

   $A,3000,82817.88,27.50,-2.46,-7.29,-4.11

The columns are: sentence type, sequence number, baro, oat, dp0, dpA,
dpB.

A battery management sentence:

   $B,75,4.07,-84.38,1545.60,0.11

The columns are: sentence type, sequence number, battery voltage,
battery charge/discharge rate (mA), battery capacity in mAh and
fraction. While the voltage and rate will remain, the rest of the
sentence is still subject to change.

The sentence type is new in this version; previous versions of the probe
software's output correspond to the airdata/$A sentence. The sequence
number is generated on the probe and is newly added in order to allow
the display software (and other analysis) to detect missing data.

Additionally the probe is now using XBee's API mode to send data,
although it is not required to do so to listen to the data stream.
(Doing so does allow receipt for RSSI signal strength information
however, and is recommended.)