Created a concrete implementation of TelemetryClient, called XbeeTelemetryClient, that refactors into a reusable component the functionality of the existing airball_probe_telemetry.cpp. This component is responsible for logging as well as acquiring data.
Wired up XbeeTelemetryClient into the display code, and removed the
legacy logging functionality from the display code.

This work has yet to be bench-tested. However, the risk of breaking
airball_probe_telemetry.cpp with these refactorings is very low. The
main display program ab_device.cpp has been broken anyway while this
work has been going on so it's not a regression. The demo program
ab.cpp is verified to be working properly with dummy data.