A previous change pushed the XBee setup logic into class "xbee". It turns out that this was a less-workable abstraction, so this change reverts that change while keeping the simple refactoring (proper C++ header/implementation, etc).
For now, that means that the setup logic (e.g. AT commands) is
duplicated in a bunch of places. We need some abstraction to clean that
up, since we're going to have to support the 900 MHz XBee modules which
have rather different behavior from the 2.4 GHz modules we're using now,
and we don't want to write any more duplication. We will work on that
abstraction next.

The state as of this CL is tested to work with the 2.4 GHz XBee modules,
so there should be no regression.