Added new frame types for sending and receiving AT commands in API mode.
The problem this change was trying to solve is that, if you are hoping
to configure the XBee based on data you *read* from it, and it's already
in transparent mode and receiving data, then you may end up eating a lot
of garbage data and it's hard to synchronize.

My attempted solution was to try to *always* go into API mode, and then
send everything, including AT commands, in API mode. That way we could
read back various status bits (e.g. the hardware type of the XBee) and
be assured of synchronizing properly in the various data streams.

Pushing this through was not as successful as I had hoped. An XBee that
was being configured "afresh" was usually okay, but one that was
previously being used would often fail to read the type correctly.

As a result, I'm checking in this change so the work I did is not
completely lost, but I'm now assuming we will be passing the current
XBee "type" as a command line parameter in all cases, or hard-coding it
where it's more appropriate to do so.

As of this change, the telemetry work does not regress for 802.14 XBees.

This change also does work (woo hoo!) for the 900 MHz XBees. It seems to
operate with somewhat different latency characteristics than we expect,
even though we are using the 200 Kbps firmware. We will examine this
further.